+++ src/app/objects/ListingFilter.java
@@ -24,11 +24,11 @@ public class ListingFilter {
         if (listing.containsKey("postalCode")) this.postalCode = (String) listing.get("postalCode");
         if (listing.containsKey("country")) this.country = (String) listing.get("country");
         if (listing.containsKey("type")) this.type = (String) listing.get("type");
-        if (listing.containsKey("longitude")) this.longitude = (BigDecimal) listing.get("longitude");
-        if (listing.containsKey("latitude")) this.latitude = (BigDecimal) listing.get("latitude");
-        if (listing.containsKey("lowestPrice")) this.lowestPrice = (BigDecimal) listing.get("lowestPrice");
-        if (listing.containsKey("highestPrice")) this.highestPrice = (BigDecimal) listing.get("highestPrice");
-        if (listing.containsKey("maxDistance")) this.maxDistance = (BigDecimal) listing.get("maxDistance");
+        if (listing.containsKey("longitude")) this.longitude = BigDecimal.valueOf((double) listing.get("longitude"));
+        if (listing.containsKey("latitude")) this.latitude = BigDecimal.valueOf((double) listing.get("latitude"));
+        if (listing.containsKey("lowestPrice")) this.lowestPrice = BigDecimal.valueOf((double) listing.get("lowestPrice"));
+        if (listing.containsKey("highestPrice")) this.highestPrice = BigDecimal.valueOf((double) listing.get("highestPrice"));
+        if (listing.containsKey("maxDistance")) this.maxDistance = BigDecimal.valueOf((double) listing.get("maxDistance"));
         if (listing.containsKey("firstDate")) this.firstDate = (Date) listing.get("firstDate");
         if (listing.containsKey("lastDate")) this.lastDate = (Date) listing.get("lastDate");
         if (listing.containsKey("minNumberOfBedrooms")) this.minNumberOfBedrooms = (int) listing.get("minNumberOfBedrooms");
